+++ src/Strategy/ElectronicItemStrategy.php
@@ -1,9 +1,11 @@
 <?php
 declare(strict_types=1);
+// this is example class with new item category
+namespace App\Strategy;
 
-namespace App\Model\Item;
+use App\Model\Item\AbstractItem;
 
-class Legendary extends AbstractItem
+class Electronic extends AbstractItem
 {
     public function updateQuality(): void
     {
