+++ src/Strategy/LegendaryItemStrategy.php
@@ -1,9 +1,11 @@
 <?php
 declare(strict_types=1);
-// this is example class with new item category
-namespace App\Model\Item;
 
-class Electronic extends AbstractItem
+namespace App\Strategy;
+
+use App\Model\Item\AbstractItem;
+
+class Legendary extends AbstractItem
 {
     public function updateQuality(): void
     {
